Probando Test

+++ src/ar/edu/unq/po2/tpintegrador/EstacionamientoViaApp.java
@@ -6,7 +6,7 @@ public class EstacionamientoViaApp extends Estacionamiento {
 	private int numeroDeCelular;
 	
 	private String patente;
-	private Integer saldo;
+	private float saldo;
 	private Sem sem;
 	private Boolean estaActivo;
